move tests that don't require per Test cluster setup
speeds up test execution
